backend/python_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-mongoengine==1.0.0
mongoengine==0.27.0
pymongo==4.6.2
python-dotenv==1.0.0
PyJWT==2.8.0